[DEV-2268] Refactoring
