--- src/subdomains/supporting/payin/strategies/register/impl/base/jellyfish.strategy.ts
@@ -1,8 +1,0 @@
-import { PayInRepository } from 'src/subdomains/supporting/payin/repositories/payin.repository';
-import { RegisterStrategy } from './register.strategy';
-
-export abstract class JellyfishStrategy extends RegisterStrategy {
-  constructor(payInRepository: PayInRepository) {
-    super(payInRepository);
-  }
-}
